requirements.txt
requests==2.24.0
selenium==3.141.0
pandas==1.0.5
beautifulsoup4==4.9.3
Pillow==8.1.0
